// src/app/Models/clientProfile.model.ts
export class clientDetails{
    firstName: string;
    middleNames: string;
    surname: string;
    dateOfBirth: Date;
    homePhone: string;
    workPhone: string;
    mobilePhone: string;
    addressLine1: string;
    addressLine2: string;
    townCity: string;
    country: string;
    postCode: string;
    
}

export default clientDetails;